+++ src/utils/getHtmlTable.ts
@@ -1,6 +1,9 @@
-import { TestResults } from "..";
+import { TestCase } from "@playwright/test/reporter";
+import Convert from "ansi-to-html";
 
-export const getHtmlTable = (tests: TestResults): string => {
+export const getHtmlTable = (tests: TestCase[], showError: boolean): string => {
+  const convert = new Convert();
+
   const content: string[] = [];
 
   content.push(`<br>`);
@@ -10,19 +13,34 @@ export const getHtmlTable = (tests: TestResults): string => {
   content.push(`<th>Test</th>`);
   content.push(`<th>Status</th>`);
   content.push(`<th>Duration</th>`);
+  content.push(`<th>Retries</th>`);
+  if (showError) {
+    content.push(`<th>Error</th>`);
+  }
   content.push(`</tr>`);
   content.push(`</thead>`);
   content.push(`<tbody>`);
 
-  for (const testName of Object.keys(tests)) {
-    const test = tests[testName];
+  for (const test of tests) {
+    // Get the last result
+    const result = test.results[test.results.length - 1];
 
     content.push(`<tr>`);
-    content.push(`<td>${testName}</td>`);
+    content.push(`<td>${test.title}</td>`);
     content.push(
-      `<td>${test.status === "passed" ? "✅ Pass" : "❌ Fail"}</td>`
+      `<td>${result.status === "passed" ? "✅ Pass" : "❌ Fail"}</td>`
     );
-    content.push(`<td>${test.duration / 1000}s</td>`);
+    content.push(`<td>${result.duration / 1000}s</td>`);
+    content.push(`<td>${result.retry}</td>`);
+    if (showError) {
+      content.push(
+        `<td>${
+          result.error && result.error.message
+            ? convert.toHtml(result.error.message!)
+            : ""
+        }</td>`
+      );
+    }
     content.push(`</tr>`);
   }