+++ src/utils/getTableRows.ts
@@ -1,41 +1,75 @@
 import { SummaryTableRow } from "@actions/core/lib/summary";
-import { TestResults } from "..";
+import { TestCase } from "@playwright/test/reporter";
+import Convert from "ansi-to-html";
 
-export const getTableRows = (tests: TestResults): SummaryTableRow[] => {
-  const tableRows = [
-    [
-      {
-        data: "Test",
-        header: true,
-      },
-      {
-        data: "Status",
-        header: true,
-      },
-      {
-        data: "Duration",
-        header: true,
-      },
-    ],
+export const getTableRows = (
+  tests: TestCase[],
+  showError: boolean
+): SummaryTableRow[] => {
+  const convert = new Convert();
+
+  const tableHeaders = [
+    {
+      data: "Test",
+      header: true,
+    },
+    {
+      data: "Status",
+      header: true,
+    },
+    {
+      data: "Duration",
+      header: true,
+    },
+    {
+      data: "Retries",
+      header: true,
+    },
   ];
 
-  for (const testName of Object.keys(tests)) {
-    const test = tests[testName];
+  if (showError) {
+    tableHeaders.push({
+      data: "Error",
+      header: true,
+    });
+  }
 
-    tableRows.push([
+  const tableRows = [tableHeaders];
+
+  for (const test of tests) {
+    // Get the last result
+    const result = test.results[test.results.length - 1];
+
+    const tableRow = [
       {
-        data: testName,
+        data: test.title,
         header: false,
       },
       {
-        data: test.status === "passed" ? "✅ Pass" : "❌ Fail",
+        data: result.status === "passed" ? "✅ Pass" : "❌ Fail",
         header: false,
       },
       {
-        data: `${test.duration / 1000}s`,
+        data: `${result.duration / 1000}s`,
         header: false,
       },
-    ]);
+      {
+        data: `${result.retry}`,
+        header: false,
+      },
+    ];
+
+    if (showError) {
+      tableRow.push({
+        data:
+          result.error && result.error.message
+            ? convert.toHtml(result.error.message!)
+            : "",
+        header: false,
+      });
+    }
+
+    tableRows.push(tableRow);
   }
 
   return tableRows;