+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0]
+
+- Update test suite logic
+- Added OS name
+- Added browser/project name
+- Added retries count
+- Added `showError` option
+
 ## [1.0.0]
 
 - Initial release